Windows support for meds2pdb NB: database schema changed: virtual_address_offset are now 64bit
Former-commit-id: 8a965317cc568b9d662a93d57b72c26117be56b8

+++ libEXEIO/include/exeio.h
@@ -14,8 +14,7 @@ namespace EXEIO
 
 	typedef enum { ELF64, ELF32, PE32, PE64 } execlass_t;
 
-	typedef int virtual_offset_t;
-
+	typedef uintptr_t virtual_offset_t;
 
 
 	// create type to match elfio
@@ -76,7 +75,7 @@ namespace EXEIO
 			exeio_t(char* filename) { Init(); load(filename); }
 			~exeio_t() { delete backend; }
 
-                        virtual void load(std::string filename) { load((char*)filename.c_str()); }
+            virtual void load(std::string filename) { load((char*)filename.c_str()); }
 
 			// load the file
 			virtual void load(char* fn);
